+++ packages/rocketh-deploy-proxy/src/index.ts
@@ -145,7 +145,11 @@ extendEnvironment((env: Environment) => {
 
 			const implementationSlotData = await env.network.provider.request({
 				method: 'eth_getStorageAt',
-				params: [proxyDeployment.address, '0x360894a13ba1a3210667c828492db98dca3e2076cc3735a920a3ca505d382bbc'],
+				params: [
+					proxyDeployment.address,
+					'0x360894a13ba1a3210667c828492db98dca3e2076cc3735a920a3ca505d382bbc',
+					'latest',
+				],
 			});
 			const currentImplementationAddress = `0x${implementationSlotData.substr(-40)}`;