+++ packages/rocketh-verifier/src/cli.ts
@@ -5,6 +5,7 @@ import {run} from '.';
 import {Command, Option} from 'commander';
 import pkg from '../package.json';
 import {ConfigOptions} from 'rocketh';
+import {exportMetadata} from './metadata';
 loadEnv();
 
 const commandName = `rocketh-verify`;
@@ -38,6 +39,16 @@ program
 		const programOptions = program.opts() as ConfigOptions;
 		const resolvedConfig = readAndResolveConfig(programOptions, {ignoreMissingRPC: true});
 		run(resolvedConfig, {verifier: {type: 'sourcify', endpoint: options.endpoint}});
+	});
+
+program
+	.command('metadata')
+	.description('export metadata')
+	.option('--out <value>', 'folder to write metadata into')
+	.action((options: {out?: string}) => {
+		const programOptions = program.opts() as ConfigOptions;
+		const resolvedConfig = readAndResolveConfig(programOptions, {ignoreMissingRPC: true});
+		exportMetadata(resolvedConfig, {out: options.out || '_metadata'});
 	});
 
 program.parse(process.argv);